+++ packages/tautulli/lib/src/commands/activity/delete_temp_sessions.dart
@@ -1,9 +1,9 @@
 part of tautulli_commands;
 
-Future<void> _commandRestart(Dio client) async {
+Future<void> _commandDeleteTempSessions(Dio client) async {
     Response response = await client.get('/',
         queryParameters: {
-            'cmd': 'restart',
+            'cmd': 'delete_temp_sessions',
         },
     );
     switch((response.data['response']['result'] as String)) {