+++ packages/tautulli/lib/src/commands/libraries/delete_recently_added.dart
@@ -1,9 +1,9 @@
 part of tautulli_commands;
 
-Future<void> _commandRestart(Dio client) async {
+Future<void> _commandDeleteRecentlyAdded(Dio client) async {
     Response response = await client.get('/',
         queryParameters: {
-            'cmd': 'restart',
+            'cmd': 'delete_recently_added',
         },
     );
     switch((response.data['response']['result'] as String)) {