+++ packages/tautulli/lib/src/commands/libraries/refresh_libraries_list.dart
@@ -9,6 +9,6 @@ Future<void> _commandRefreshLibrariesList(Dio client) async {
     switch((response.data['response']['result'] as String)) {
         case 'success': return;
         case 'error':
-        default: throw Exception(throw Exception(response.data['response']['message']));
+        default: throw Exception(response.data['response']['message']);
     }
 }