+++ packages/tautulli/lib/src/commands/notifications/add_notifier_config.dart
@@ -13,6 +13,6 @@ Future<String> _commandAddNotifierConfig(Dio client, {
     switch((response.data['response']['result'] as String)) {
         case 'success': return response.data['response']['data'];
         case 'error':
-        default: throw Exception(throw Exception(response.data['response']['message']));
+        default: throw Exception(response.data['response']['message']);
     }
 }