+++ packages/tautulli/lib/src/commands/system/delete_media_info_cache.dart
@@ -3,6 +3,7 @@ part of tautulli_commands;
 Future<void> _commandDeleteMediaInfoCache(Dio client, {
     @required int sectionId,
 }) async {
+    assert(sectionId != null, 'sectionId cannot be null.');
     Response response = await client.get('/',
         queryParameters: {
             'cmd': 'delete_media_info_cache',
@@ -12,6 +13,6 @@ Future<void> _commandDeleteMediaInfoCache(Dio client, {
     switch((response.data['response']['result'] as String)) {
         case 'success': return;
         case 'error':
-        default: throw Exception(throw Exception(response.data['response']['message']));
+        default: throw Exception(response.data['response']['message']);
     }
 }