+++ packages/tautulli/lib/src/commands/system/status.dart
@@ -10,6 +10,6 @@ Future<String> _commandStatus(Dio client) async {
     switch((response.data['response']['result'] as String)) {
         case 'success': return response.data['response']['message'];
         case 'error':
-        default: throw Exception(throw Exception(response.data['response']['message']));
+        default: throw Exception(response.data['response']['message']);
     }
 }