+++ packages/tautulli/lib/src/commands/users/delete_all_user_history.dart
@@ -15,6 +15,6 @@ Future<void> _commandDeleteAllUserHistory(Dio client, {
     switch((response.data['response']['result'] as String)) {
         case 'success': return;
         case 'error':
-        default: throw Exception(throw Exception(response.data['response']['message']));
+        default: throw Exception(response.data['response']['message']);
     }
 }