+++ packages/tautulli/lib/src/commands/users/delete_user.dart
@@ -1,20 +1,18 @@
 part of tautulli_commands;
 
-Future<void> _commandDeleteLibrary(Dio client, {
-    @required int sectionId,
-    String serverId,
+Future<void> _commandDeleteUser(Dio client, {
+    @required int userId,
     List<int> rowIds,
 }) async {
-    assert(sectionId != null, 'sectionId cannot be null.');
+    assert(userId != null, 'userId cannot be null.');
     Response response = await client.get('/',
         queryParameters: {
-            'cmd': 'delete_library',
-            'section_id': sectionId,
-            if(serverId != null) 'server_id': serverId,
+            'cmd': 'delete_user',
+            'user_id': userId,
             if(rowIds != null && rowIds.isNotEmpty) 'row_ids': rowIds.join(","),
         },
     );
-    switch((response.data['response']['result'] as String)) {
+    switch(response.data['response']['result']) {
         case 'success': return;
         case 'error':
         default: throw Exception(response.data['response']['message']);