+++ packages/tautulli/lib/src/types/image_host_service.dart
@@ -1,6 +1,6 @@
 part of tautulli_types;
 
-/// Enumerator to handle all image hosting services
+/// Enumerator to handle all image hosting services available in Tautulli.
 enum TautulliImageHostService {
     IMGUR,
     CLOUDINARY,